% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_InfoAnnot.R
\name{getModelsList}
\alias{getModelsList}
\title{List the models available}
\usage{
getModelsList()
}
\value{
data.frame of all available models to use with specified assembly and population
}
\description{
This function prints the list of all available models.
}
