% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_InfoAnnot.R
\name{getSamplesInfo}
\alias{getSamplesInfo}
\title{List the samples annotation}
\usage{
getSamplesInfo()
}
\description{
This function prints the list of all 1,000 Genomes Project samples used to build the reference models.
}
